# frontend/openapi.json
{"openapi":"3.0.3","info":{"title":"fleet_management_backend","description":"","license":{"name":""},"version":"0.1.0"},"paths":{"/api/auth/login":{"post":{"tags":["crate::routes::auth"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}},"required":true},"responses":{"200":{"description":"User logged in successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"401":{"description":"Unauthorized"}}}},"/api/auth/me":{"get":{"tags":["crate::routes::auth"],"operationId":"me","responses":{"200":{"description":"Current user info","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Claims"}}}},"401":{"description":"Unauthorized"}},"security":[{"jwt":[]}]}},"/api/auth/register":{"post":{"tags":["crate::routes::auth"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}},"required":true},"responses":{"200":{"description":"User registered successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"500":{"description":"Internal Server Error"}}}}},"components":{"schemas":{"Alert":{"type":"object","required":["id","entity_id","type","severity","is_resolved","created_at"],"properties":{"created_at":{"type":"string","format":"date-time"},"entity_id":{"type":"string","format":"uuid"},"id":{"type":"string","format":"uuid"},"is_resolved":{"type":"boolean"},"resolved_at":{"type":"string","format":"date-time","nullable":true},"severity":{"$ref":"#/components/schemas/AlertSeverity"},"type":{"type":"string"}}},"AlertSeverity":{"type":"string","enum":["Low","Medium","High","Critical"]},"AssignmentStatus":{"type":"string","enum":["Scheduled","Active","Completed","Cancelled"]},"AuthResponse":{"type":"object","required":["token","user"],"properties":{"token":{"type":"string"},"user":{"$ref":"#/components/schemas/User"}}},"Claims":{"type":"object","required":["sub","role","exp"],"properties":{"exp":{"type":"integer","minimum":0},"role":{"$ref":"#/components/schemas/UserRole"},"sub":{"type":"string","format":"uuid"}}},"CreateAlertDto":{"type":"object","required":["entity_id","type","severity"],"properties":{"entity_id":{"type":"string","format":"uuid"},"severity":{"$ref":"#/components/schemas/AlertSeverity"},"type":{"type":"string"}}},"CreateAssignmentDto":{"type":"object","required":["vehicle_id","driver_id","start_time","status"],"properties":{"driver_id":{"type":"string","format":"uuid"},"end_time":{"type":"string","format":"date-time","nullable":true},"start_time":{"type":"string","format":"date-time"},"status":{"$ref":"#/components/schemas/AssignmentStatus"},"vehicle_id":{"type":"string","format":"uuid"}}},"CreateCustomerDto":{"type":"object","required":["name","contact_info","billing_address"],"properties":{"billing_address":{"type":"string"},"contact_info":{"type":"object"},"name":{"type":"string"}}},"CreateDriverDto":{"type":"object","required":["user_id","license_number","status"],"properties":{"license_number":{"type":"string"},"status":{"$ref":"#/components/schemas/DriverStatus"},"user_id":{"type":"string","format":"uuid"}}},"CreateMaintenanceRecordDto":{"type":"object","required":["vehicle_id","type","cost","date","provider"],"properties":{"cost":{"type":"string"},"date":{"type":"string","format":"date-time"},"description":{"type":"string","nullable":true},"provider":{"type":"string"},"type":{"$ref":"#/components/schemas/MaintenanceType"},"vehicle_id":{"type":"string","format":"uuid"}}},"CreateRouteDto":{"type":"object","required":["job_id","origin","destination","waypoints"],"properties":{"destination":{"type":"object"},"job_id":{"type":"string","format":"uuid"},"origin":{"type":"object"},"waypoints":{"type":"object"}}},"CreateShipmentDto":{"type":"object","required":["job_id","weight","dimensions","type"],"properties":{"dimensions":{"type":"object"},"job_id":{"type":"string","format":"uuid"},"type":{"type":"string"},"weight":{"type":"number","format":"double"}}},"CreateTransportJobDto":{"type":"object","required":["customer_id","status","agreed_price"],"properties":{"agreed_price":{"type":"string"},"customer_id":{"type":"string","format":"uuid"},"status":{"$ref":"#/components/schemas/JobStatus"}}},"CreateUserDto":{"type":"object","required":["email","password_hash","role","is_active"],"properties":{"email":{"type":"string"},"is_active":{"type":"boolean"},"password_hash":{"type":"string"},"role":{"$ref":"#/components/schemas/UserRole"}}},"CreateVehicleDto":{"type":"object","required":["make","model","year","vin","license_plate","type","current_mileage","fuel_type"],"properties":{"current_mileage":{"type":"integer","format":"int32"},"fuel_type":{"$ref":"#/components/schemas/FuelType"},"license_plate":{"type":"string"},"make":{"type":"string"},"model":{"type":"string"},"specs":{"nullable":true},"type":{"$ref":"#/components/schemas/VehicleType"},"vin":{"type":"string"},"year":{"type":"integer","format":"int32"}}},"CreateVehicleTelemetryDto":{"type":"object","required":["time","vehicle_id","location","speed","fuel_level","engine_status"],"properties":{"engine_status":{"type":"object"},"fuel_level":{"type":"number","format":"double"},"location":{"type":"object"},"speed":{"type":"number","format":"double"},"time":{"type":"string","format":"date-time"},"vehicle_id":{"type":"string","format":"uuid"}}},"Customer":{"type":"object","required":["id","name","contact_info","billing_address","created_at","updated_at"],"properties":{"billing_address":{"type":"string"},"contact_info":{"type":"object"},"created_at":{"type":"string","format":"date-time"},"deleted_at":{"type":"string","format":"date-time","nullable":true},"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"updated_at":{"type":"string","format":"date-time"}}},"Driver":{"type":"object","required":["id","user_id","license_number","status","created_at","updated_at"],"properties":{"created_at":{"type":"string","format":"date-time"},"deleted_at":{"type":"string","format":"date-time","nullable":true},"id":{"type":"string","format":"uuid"},"license_number":{"type":"string"},"status":{"$ref":"#/components/schemas/DriverStatus"},"updated_at":{"type":"string","format":"date-time"},"user_id":{"type":"string","format":"uuid"}}},"DriverStatus":{"type":"string","enum":["Available","OnDuty","OffDuty","SickLeave"]},"FuelType":{"type":"string","enum":["Diesel","Gasoline","Electric"]},"JobStatus":{"type":"string","enum":["Pending","InProgress","Delivered","Invoiced","Paid"]},"LoginDto":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string"},"password":{"type":"string"}}},"MaintenanceRecord":{"type":"object","required":["id","vehicle_id","type","cost","date","provider","created_at"],"properties":{"cost":{"type":"string"},"created_at":{"type":"string","format":"date-time"},"date":{"type":"string","format":"date-time"},"description":{"type":"string","nullable":true},"id":{"type":"string","format":"uuid"},"provider":{"type":"string"},"type":{"$ref":"#/components/schemas/MaintenanceType"},"vehicle_id":{"type":"string","format":"uuid"}}},"MaintenanceType":{"type":"string","enum":["Preventive","Repair","Inspection","Accident"]},"MonthlyFinancialSummary":{"type":"object","required":["month","revenue","cost","profit"],"properties":{"cost":{"type":"string"},"month":{"type":"string"},"profit":{"type":"string"},"revenue":{"type":"string"}}},"Role":{"type":"object","required":["id","name","permissions"],"properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"permissions":{}}},"Route":{"type":"object","required":["id","job_id","origin","destination","waypoints"],"properties":{"destination":{"type":"object"},"id":{"type":"string","format":"uuid"},"job_id":{"type":"string","format":"uuid"},"origin":{"type":"object"},"waypoints":{"type":"object"}}},"Shipment":{"type":"object","required":["id","job_id","weight","dimensions","type"],"properties":{"dimensions":{"type":"object"},"id":{"type":"string","format":"uuid"},"job_id":{"type":"string","format":"uuid"},"type":{"type":"string"},"weight":{"type":"number","format":"double"}}},"TransportJob":{"type":"object","required":["id","customer_id","status","agreed_price","created_at","updated_at"],"properties":{"agreed_price":{"type":"string"},"created_at":{"type":"string","format":"date-time"},"customer_id":{"type":"string","format":"uuid"},"id":{"type":"string","format":"uuid"},"status":{"$ref":"#/components/schemas/JobStatus"},"updated_at":{"type":"string","format":"date-time"}}},"User":{"type":"object","required":["id","email","role","is_active","created_at","updated_at"],"properties":{"created_at":{"type":"string","format":"date-time"},"deleted_at":{"type":"string","format":"date-time","nullable":true},"email":{"type":"string"},"id":{"type":"string","format":"uuid"},"is_active":{"type":"boolean"},"role":{"$ref":"#/components/schemas/UserRole"},"updated_at":{"type":"string","format":"date-time"}}},"UserRole":{"type":"string","enum":["Admin","Manager","Driver","Mechanic"]},"Vehicle":{"type":"object","required":["id","make","model","year","vin","license_plate","type","status","current_mileage","fuel_type","created_at","updated_at"],"properties":{"created_at":{"type":"string","format":"date-time"},"current_mileage":{"type":"integer","format":"int32"},"deleted_at":{"type":"string","format":"date-time","nullable":true},"fuel_type":{"$ref":"#/components/schemas/FuelType"},"id":{"type":"string","format":"uuid"},"license_plate":{"type":"string"},"make":{"type":"string"},"model":{"type":"string"},"specs":{"nullable":true},"status":{"$ref":"#/components/schemas/VehicleStatus"},"type":{"$ref":"#/components/schemas/VehicleType"},"updated_at":{"type":"string","format":"date-time"},"vin":{"type":"string"},"year":{"type":"integer","format":"int32"}}},"VehicleAssignment":{"type":"object","required":["id","vehicle_id","driver_id","start_time","status","created_at","updated_at"],"properties":{"created_at":{"type":"string","format":"date-time"},"driver_id":{"type":"string","format":"uuid"},"end_time":{"type":"string","format":"date-time","nullable":true},"id":{"type":"string","format":"uuid"},"start_time":{"type":"string","format":"date-time"},"status":{"$ref":"#/components/schemas/AssignmentStatus"},"updated_at":{"type":"string","format":"date-time"},"vehicle_id":{"type":"string","format":"uuid"}}},"VehicleProfitability":{"type":"object","required":["vehicle_id","vehicle_plate","revenue","cost","profit","rank"],"properties":{"cost":{"type":"string"},"profit":{"type":"string"},"rank":{"type":"integer","format":"int32"},"revenue":{"type":"string"},"vehicle_id":{"type":"string","format":"uuid"},"vehicle_plate":{"type":"string"}}},"VehicleStatus":{"type":"string","enum":["Available","Assigned","Maintenance","OutOfService"]},"VehicleTelemetry":{"type":"object","required":["time","vehicle_id","location","speed","fuel_level","engine_status"],"properties":{"engine_status":{"type":"object"},"fuel_level":{"type":"number","format":"double"},"location":{"type":"object"},"speed":{"type":"number","format":"double"},"time":{"type":"string","format":"date-time"},"vehicle_id":{"type":"string","format":"uuid"}}},"VehicleType":{"type":"string","enum":["Truck","Van","Sedan"]}},"securitySchemes":{"jwt":{"type":"http","scheme":"bearer","bearerFormat":"JWT","description":"JWT Authorization header using the Bearer scheme."}}},"tags":[{"name":"fleet-management","description":"Fleet Management System API"}]}